matching-algorithm/requirements.txt
pandas
psycopg2-binary
sqlalchemy
flask
flask_cors
waitress
numpy